# backend/openapi.json
{"openapi":"3.1.0","info":{"title":"FastAPI","version":"0.1.0"},"paths":{"/":{"get":{"summary":"Index","operationId":"index__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/v1/auth/register":{"post":{"tags":["auth"],"summary":"Register User","operationId":"register_user_api_v1_auth_register_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserCreate"}}},"required":true},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/auth/token":{"post":{"tags":["auth"],"summary":"Token","operationId":"token_api_v1_auth_token_post","requestBody":{"content":{"application/x-www-form-urlencoded":{"schema":{"$ref":"#/components/schemas/Body_token_api_v1_auth_token_post"}}},"required":true},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Token"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/auth/logout":{"post":{"tags":["auth"],"summary":"Logout","operationId":"logout_api_v1_auth_logout_post","responses":{"204":{"description":"Successful Response"}},"security":[{"OAuth2PasswordBearer":[]}]}},"/api/v1/auth/me":{"get":{"tags":["auth"],"summary":"Read Me","operationId":"read_me_api_v1_auth_me_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserOut"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/news/subscribe":{"get":{"tags":["news"],"summary":"Get Subscription Status","operationId":"get_subscription_status_news_subscribe_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}},"security":[{"OAuth2PasswordBearer":[]}]},"post":{"tags":["news"],"summary":"Subscribe To News","operationId":"subscribe_to_news_news_subscribe_post","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}},"security":[{"OAuth2PasswordBearer":[]}]},"delete":{"tags":["news"],"summary":"Unsubscribe From News","operationId":"unsubscribe_from_news_news_subscribe_delete","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/news/":{"get":{"tags":["news"],"summary":"Get News Endpoint","operationId":"get_news_endpoint_news__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/NewsResponse"},"type":"array","title":"Response Get News Endpoint News  Get"}}}}},"security":[{"OAuth2PasswordBearer":[]}]},"post":{"tags":["news"],"summary":"Create News Endpoint","operationId":"create_news_endpoint_news__post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/NewsCreate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/NewsResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/news/{news_id}":{"get":{"tags":["news"],"summary":"Get Single News","operationId":"get_single_news_news__news_id__get","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"news_id","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"News Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/NewsResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"tags":["news"],"summary":"Delete News Endpoint","operationId":"delete_news_endpoint_news__news_id__delete","parameters":[{"name":"news_id","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"News Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/news/{news_id}/like":{"post":{"tags":["news"],"summary":"Like News Endpoint","operationId":"like_news_endpoint_news__news_id__like_post","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"news_id","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"News Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"tags":["news"],"summary":"Unlike News Endpoint","operationId":"unlike_news_endpoint_news__news_id__like_delete","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"news_id","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"News Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/news/{news_id}/view":{"post":{"tags":["news"],"summary":"View News Endpoint","operationId":"view_news_endpoint_news__news_id__view_post","parameters":[{"name":"news_id","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"News Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/courses/":{"get":{"tags":["courses"],"summary":"List Courses","operationId":"list_courses_courses__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/CourseListItem"},"type":"array","title":"Response List Courses Courses  Get"}}}}},"security":[{"OAuth2PasswordBearer":[]}]},"post":{"tags":["courses"],"summary":"Create Course Endpoint","operationId":"create_course_endpoint_courses__post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CourseCreate"}}},"required":true},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CourseResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/courses/{course_id}":{"get":{"tags":["courses"],"summary":"Get Course","operationId":"get_course_courses__course_id__get","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"course_id","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"Course Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CourseResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/courses/{course_id}/materials":{"post":{"tags":["courses"],"summary":"Add Course Material","operationId":"add_course_material_courses__course_id__materials_post","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"course_id","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"Course Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CourseMaterialCreate"}}}},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CourseMaterialResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/courses/{course_id}/subscribe":{"post":{"tags":["courses"],"summary":"Subscribe","operationId":"subscribe_courses__course_id__subscribe_post","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"course_id","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"Course Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"tags":["courses"],"summary":"Unsubscribe","operationId":"unsubscribe_courses__course_id__subscribe_delete","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"course_id","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"Course Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/metrics":{"get":{"summary":"Metrics","operationId":"metrics_metrics_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/v1/telegram/link-token":{"post":{"tags":["telegram"],"summary":"Generate Link Token","operationId":"generate_link_token_api_v1_telegram_link_token_post","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/api/v1/telegram/unlink":{"delete":{"tags":["telegram"],"summary":"Unlink Telegram","operationId":"unlink_telegram_api_v1_telegram_unlink_delete","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/api/v1/telegram/status":{"get":{"tags":["telegram"],"summary":"Telegram Status","operationId":"telegram_status_api_v1_telegram_status_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}},"security":[{"OAuth2PasswordBearer":[]}]}}},"components":{"schemas":{"AuthorShort":{"properties":{"id":{"type":"string","format":"uuid","title":"Id"},"first_name":{"type":"string","title":"First Name"},"last_name":{"type":"string","title":"Last Name"}},"type":"object","required":["id","first_name","last_name"],"title":"AuthorShort"},"Body_token_api_v1_auth_token_post":{"properties":{"grant_type":{"type":"string","pattern":"password|refresh_token","title":"Grant Type"},"email":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Email"},"password":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Password"},"refresh_token":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Refresh Token"}},"type":"object","required":["grant_type"],"title":"Body_token_api_v1_auth_token_post"},"CourseCreate":{"properties":{"name":{"type":"string","title":"Name"},"description":{"type":"string","title":"Description"}},"type":"object","required":["name","description"],"title":"CourseCreate"},"CourseListItem":{"properties":{"id":{"type":"string","format":"uuid","title":"Id"},"name":{"type":"string","title":"Name"},"description":{"type":"string","title":"Description"},"teacher_id":{"type":"string","format":"uuid","title":"Teacher Id"},"created_at":{"type":"string","format":"date-time","title":"Created At"},"material_count":{"type":"integer","title":"Material Count","default":0},"is_subscribed":{"type":"boolean","title":"Is Subscribed","default":false}},"type":"object","required":["id","name","description","teacher_id","created_at"],"title":"CourseListItem"},"CourseMaterialCreate":{"properties":{"title":{"type":"string","title":"Title"},"description":{"type":"string","title":"Description"},"material_type":{"type":"string","title":"Material Type","default":"material"}},"type":"object","required":["title","description"],"title":"CourseMaterialCreate"},"CourseMaterialResponse":{"properties":{"id":{"type":"string","format":"uuid","title":"Id"},"course_id":{"type":"string","format":"uuid","title":"Course Id"},"title":{"type":"string","title":"Title"},"description":{"type":"string","title":"Description"},"material_type":{"type":"string","title":"Material Type"},"created_at":{"type":"string","format":"date-time","title":"Created At"}},"type":"object","required":["id","course_id","title","description","material_type","created_at"],"title":"CourseMaterialResponse"},"CourseResponse":{"properties":{"id":{"type":"string","format":"uuid","title":"Id"},"name":{"type":"string","title":"Name"},"description":{"type":"string","title":"Description"},"teacher_id":{"type":"string","format":"uuid","title":"Teacher Id"},"created_at":{"type":"string","format":"date-time","title":"Created At"},"materials":{"items":{"$ref":"#/components/schemas/CourseMaterialResponse"},"type":"array","title":"Materials","default":[]},"is_subscribed":{"type":"boolean","title":"Is Subscribed","default":false}},"type":"object","required":["id","name","description","teacher_id","created_at"],"title":"CourseResponse"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"NewsCreate":{"properties":{"title":{"type":"string","title":"Title"},"category":{"type":"string","title":"Category"},"excerpt":{"type":"string","title":"Excerpt"},"content":{"type":"string","title":"Content"}},"type":"object","required":["title","category","excerpt","content"],"title":"NewsCreate"},"NewsResponse":{"properties":{"id":{"type":"string","format":"uuid","title":"Id"},"title":{"type":"string","title":"Title"},"category":{"type":"string","title":"Category"},"excerpt":{"type":"string","title":"Excerpt"},"content":{"type":"string","title":"Content"},"author_id":{"type":"string","format":"uuid","title":"Author Id"},"created_at":{"type":"string","format":"date-time","title":"Created At"},"author":{"$ref":"#/components/schemas/AuthorShort"},"likes_count":{"type":"integer","title":"Likes Count"},"is_liked":{"type":"boolean","title":"Is Liked"}},"type":"object","required":["id","title","category","excerpt","content","author_id","created_at","author","likes_count","is_liked"],"title":"NewsResponse"},"Token":{"properties":{"access_token":{"type":"string","title":"Access Token"},"refresh_token":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Refresh Token"},"token_type":{"type":"string","title":"Token Type","default":"bearer"}},"type":"object","required":["access_token"],"title":"Token"},"UserCreate":{"properties":{"email":{"type":"string","format":"email","title":"Email"},"first_name":{"type":"string","title":"First Name"},"last_name":{"type":"string","title":"Last Name"},"password":{"type":"string","title":"Password"}},"type":"object","required":["email","first_name","last_name","password"],"title":"UserCreate"},"UserOut":{"properties":{"email":{"type":"string","format":"email","title":"Email"},"first_name":{"type":"string","title":"First Name"},"last_name":{"type":"string","title":"Last Name"},"is_admin":{"type":"boolean","title":"Is Admin"},"id":{"type":"string","format":"uuid","title":"Id"}},"type":"object","required":["email","first_name","last_name","is_admin","id"],"title":"UserOut"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"},"input":{"title":"Input"},"ctx":{"type":"object","title":"Context"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}},"securitySchemes":{"OAuth2PasswordBearer":{"type":"oauth2","flows":{"password":{"scopes":{},"tokenUrl":"/api/v1/auth/token"}}}}}}